JCSP-381: review comments fixes
