+++ custom-vu/src/main/kotlin/jces1209/vu/page/AbstractIssuePage.kt
@@ -6,5 +6,5 @@ interface AbstractIssuePage {
     fun comment(): Commenting
     fun editDescription(description: String): AbstractIssuePage
     fun linkIssue(): IssueLinking
-    fun addAttachment()
+    fun addAttachment(): AttachScreenShot
 }
