+++ custom-vu/src/main/kotlin/jces1209/vu/page/CloudAddScreenShot.kt
@@ -2,7 +2,7 @@ package jces1209.vu.page
 
 import org.openqa.selenium.WebDriver
 
-class DcAddScreenShot(
+class CloudAddScreenShot(
     driver: WebDriver
 ) : AttachScreenShot(driver) {
